+++ R/apply_niche.R
@@ -14,7 +14,7 @@ apply_niche = function(x, niche_def, gc){
   #' @returns for a numeric vector input, returns a numeric vector, timing/location of events (e.g. fossil ages/locations) preserved after the niche model is applied. For a `pre_paleoTS` object as input, returns a `pre_paleoTS` object with specimens removed according to the niche model. For a `fossils` object, returns a `fossils` object with some occurrences removed according to the niche definition
   #'
   #' @seealso
-    #' * [snd_niche()] and [bounded_niche()] for template niche models, [discrete_niche()] and [discrete_gradient()] to construct niches from discrete categories.
+    #' * [snd_niche()] and [bounded_niche()] for template niche models, [discrete_niche()] and [discrete_gradient()] to construct niches from discrete categories, [trivial_niche()] to model organisms without niche specifications
     #' * `vignette("advanced_functionality)` for how to create user-defined niche models
     #' * [apply_taphonomy()] to model taphonomic effects based on a similar principle
     #' * [thin()] and [prob_remove()] for the underlying mathematical procedures