+++ tests/testthat/test_apply_niche.R
@@ -29,3 +29,12 @@ test_that("for fossils objects, returns indentical result for trivial niche",{
   expect_equal(apply_niche(f, niche_def, gc), f)
 })
 
+test_that("warning is thrown for age uncertainty",{
+  niche_def = function(x) rep(1, length(x))
+  gc = function(x) rep(1, length(x))
+  t = ape::rlineage(birth = 1, death = 0, Tmax = 1)
+  rate = 2
+  f = FossilSim::sim.fossils.poisson(rate = rate, tree = t)
+  f$hmin = f$hmin + 0.1
+  expect_warning(apply_niche(f, niche_def, gc))
+})