+++ tests/testthat/test_plot.pre_paleoTS.R
@@ -1,0 +1,4 @@
+test_that("throws error",{
+  x = stasis_sl(1:3)
+  expect_error(plot(x))
+})