+++ tests/testthat/test_trivial_niche.R
@@ -1,0 +1,3 @@
+test_that("return vector is correct",{
+  expect_equal(trivial_niche(c(2,3)), c(1,1))
+})